FEAT: Local search for SWAP-1 operation.
local_search: Goes over the path with the worst remaining time available, and swap a pair of nodes within it to increase it.
swap1: Swaps two nodes within the path and checks if the new path is feasible both in the affected nodes and the rest of the path.
get_arrival_time: New function to get the arrival time of a node in the path. Used on all the nodes within the swap.